requirements.txt
PyPDF2>=3.0.0
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.20.0
streamlit-authenticator>=0.3.2
PyGithub>=2.1.1
bcrypt>=4.1.0
PyYAML>=6.0.1
